Fix integer overflow breaking Lidgren metrics.

+++ Robust.Server/BaseServer.cs
@@ -94,8 +94,8 @@ namespace Robust.Server
         private IGameLoop _mainLoop = default!;
 
         private TimeSpan _lastTitleUpdate;
-        private int _lastReceivedBytes;
-        private int _lastSentBytes;
+        private long _lastReceivedBytes;
+        private long _lastSentBytes;
 
         private string? _shutdownReason;
